Moved PacketType class from 'enums' to 'decoder' module
These constants are only used in protocol buffer decoding.
